Revert "bulk Selector select all improvements"
This reverts commit 200d3e81e0ae2959b83dd52eec784041aa4b70d2.
