+++ src/PresentationalComponents/Inventory/Inventory.js
@@ -1,11 +1,19 @@
 import './_Inventory.scss';
 
-import { BASE_URL } from '../../AppConstants';
+import {
+  BASE_URL,
+  RULES_FETCH_URL,
+  SYSTEMS_FETCH_URL,
+} from '../../AppConstants';
 import React, { useEffect, useState } from 'react';
 import { TableVariant, sortable, wrappable } from '@patternfly/react-table';
-import { pruneFilters, urlBuilder } from '../Common/Tables';
+import {
+  pruneFilters,
+  urlBuilder,
+  workloadQueryBuilder,
+  buildTagFilter,
+} from '../Common/Tables';
 import { useDispatch, useSelector, useStore } from 'react-redux';
-import { getEntities, allCurrentSystemIds } from './helper';
 
 import DisableRule from '../../PresentationalComponents/Modals/DisableRule';
 import { Get } from '../../Utilities/Api';
@@ -14,6 +22,7 @@ import Loading from '../Loading/Loading';
 import PropTypes from 'prop-types';
 import RemediationButton from '@redhat-cloud-services/frontend-components-remediations/RemediationButton';
 import { SYSTEM_FILTER_CATEGORIES as SFC } from '../../AppConstants';
+import { mergeArraysByDiffKeys } from '../Common/Tables';
 import messages from '../../Messages';
 import { addNotification as notification } from '@redhat-cloud-services/frontend-components-notifications/';
 import { systemReducer } from '../../Store/AppReducer';
@@ -43,10 +52,8 @@ const Inventory = ({
     sort: '-last_seen',
     name: '',
   });
-  const [fullFilters, setFullFilters] = useState();
   const [total, setTotal] = useState(0);
   const entities = useSelector(({ entities }) => entities || {});
-
   const addNotification = (data) => dispatch(notification(data));
   const [disableRuleModalOpen, setDisableRuleModalOpen] = useState(false);
   const [curPageIds, setCurPageIds] = useState([]);
@@ -57,21 +64,26 @@ const Inventory = ({
   const [isRemediationButtonDisabled, setIsRemediationButtonDisabled] =
     useState(true);
 
-  const handleRefresh = (options) => {
-    /* Rec table doesn't use the same sorting params as sys table, switching between the two results in the rec table blowing up cuz its trying to
-    read the endpoint with incorrect sorting params, if we hold of on updating the sysable url params when using the this component in pathways, it
-    solves this issue for the time being*/
-    const { name, display_name } = options;
-    const refreshedFilters = {
-      ...options,
-      ...(name && {
-        name,
-      }),
-      ...(display_name && {
-        display_name,
-      }),
-    };
-    !pathway && urlBuilder(refreshedFilters, selectedTags);
+  const fetchAllSystems = async () => {
+    const allSystems = pathway
+      ? (
+          await Get(
+            `${SYSTEMS_FETCH_URL}`,
+            {},
+            {
+              pathway: pathway.slug,
+              limit: pathway.impacted_systems_count,
+            }
+          )
+        )?.data?.data?.map((system) => system.system_uuid)
+      : (
+          await Get(
+            `${RULES_FETCH_URL}${encodeURI(rule.rule_id)}/systems/`,
+            {},
+            { name: filters.name }
+          )
+        )?.data?.host_ids;
+    return allSystems;
   };
 
   const grabPageIds = () => {
@@ -86,20 +98,10 @@ const Inventory = ({
   } = useBulkSelect({
     total,
     onSelect: () => {},
-    itemIdsInTable: allCurrentSystemIds(fullFilters, total, rule),
+    itemIdsInTable: fetchAllSystems,
     itemIdsOnPage: grabPageIds,
     identitfier: 'system_uuid',
   });
-
-  const fetchSystems = getEntities(
-    handleRefresh,
-    pathway,
-    setCurPageIds,
-    setTotal,
-    selectedIds,
-    setFullFilters,
-    fullFilters
-  );
 
   // Ensures rows are marked as selected
   useEffect(() => {
@@ -274,6 +276,22 @@ const Inventory = ({
     return [displayName, tags, systemProfile, lastSeenColumn];
   };
 
+  const handleRefresh = (options) => {
+    /* Rec table doesn't use the same sorting params as sys table, switching between the two results in the rec table blowing up cuz its trying to
+    read the endpoint with incorrect sorting params, if we hold of on updating the sysable url params when using the this component in pathways, it
+    solves this issue for the time being*/
+    const { name, display_name } = options;
+    const refreshedFilters = {
+      ...options,
+      ...(name && {
+        name,
+      }),
+      ...(display_name && {
+        display_name,
+      }),
+    };
+    !pathway && urlBuilder(refreshedFilters, selectedTags);
+  };
   const removeFilterParam = (param) => {
     const filter = { ...filters, offset: 0 };
     delete filter[param];
@@ -369,7 +387,90 @@ const Inventory = ({
           SID,
         }}
         showTags={showTags}
-        getEntities={fetchSystems}
+        getEntities={async (_items, config, showTags, defaultGetEntities) => {
+          const {
+            per_page,
+            page,
+            orderBy,
+            orderDirection,
+            advisorFilters,
+            filters,
+            workloads,
+            SID,
+          } = config;
+          const sort = `${orderDirection === 'ASC' ? '' : '-'}${
+            orderBy === 'updated' ? 'last_seen' : orderBy
+          }`;
+          let options = {
+            ...advisorFilters,
+            limit: per_page,
+            offset: page * per_page - per_page,
+            sort,
+            ...(config.filters.hostnameOrId &&
+              !pathway && {
+                name: config?.filters?.hostnameOrId,
+              }),
+            ...(config.filters.hostnameOrId &&
+              pathway && {
+                display_name: config?.filters?.hostnameOrId,
+              }),
+            ...(Array.isArray(advisorFilters.rhel_version) && {
+              rhel_version: advisorFilters.rhel_version?.join(','),
+            }),
+            ...(filters.tagFilters?.length &&
+              buildTagFilter(filters.tagFilters)),
+          };
+
+          workloads &&
+            (options = { ...options, ...workloadQueryBuilder(workloads, SID) });
+
+          handleRefresh(options);
+
+          const fetchedSystems = pathway
+            ? (
+                await Get(
+                  `${SYSTEMS_FETCH_URL}`,
+                  {},
+                  { ...options, pathway: pathway.slug }
+                )
+              )?.data
+            : (
+                await Get(
+                  `${RULES_FETCH_URL}${encodeURI(
+                    rule.rule_id
+                  )}/systems_detail/`,
+                  {},
+                  options
+                )
+              )?.data;
+
+          const results = await defaultGetEntities(
+            fetchedSystems.data.map((system) => system.system_uuid),
+            {
+              page,
+              per_page,
+              hasItems: true,
+              fields: { system_profile: ['operating_system'] },
+            },
+            showTags
+          );
+          setCurPageIds(
+            fetchedSystems.data.map((system) => system.system_uuid)
+          );
+          setTotal(fetchedSystems.meta.count);
+          return Promise.resolve({
+            results: mergeArraysByDiffKeys(
+              fetchedSystems.data,
+              results.results
+            ).map((item) => {
+              return {
+                ...item,
+                selected: selectedIds?.includes(item.id),
+              };
+            }),
+            total: fetchedSystems.meta.count,
+          });
+        }}
         dedicatedAction={
           <RemediationButton
             key="remediation-button"
